sistema_ventas/requirements.txt
asgiref==3.10.0
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
crispy-bootstrap4==2022.1
cryptography==46.0.3
defusedxml==0.7.1
Django==4.2.7
django-allauth==0.52.0
django-crispy-forms==2.0
django-filter==23.3
idna==3.11
oauthlib==3.3.1
Pillow==10.0.1
pycparser==2.23
PyJWT==2.10.1
python3-openid==3.2.0
requests==2.32.5
requests-oauthlib==2.0.0
sqlparse==0.5.3
urllib3==2.5.0
